Uses dpkg-buildpackage for securedrop-app-code
Updates the Ansible-based build logic to use `dpkg-buildpackage`, so
that we can use substvars for Depends, to provide different values for
Trusty/Xenial dependencies.

As such, removes the "sed" implementation to substitute the depends
in-line, which was always a short-term solution anyway.

The fetch logic was updated to find the deb in a different location.